DEF0370542: Fixing order by limit X...Y queries selecting poor plan for execution (#87)
* DEF0370542: Fixing order by limit X...Y queries selecting high cost plan over a low cost plan.

The main question we need to answer is why optimizer is choosing a higher cost index over a lower cost index. The optimizer tracer shows that the cost of the active index is less compared to the sys_class_name2, yet the optimizer choses the sys_class_name2 index.

The reason for that is the condition here -- https://code.devsnc.com/dev/sumo-mariadb/blob/10.4.20samurai/sql/sql_select.cc#L28507
Specifically the following operation:
"keyinfo->user_defined_key_parts < best_key_parts"

For both active and sys_class_name2 we have 3 keyparts. When keyparts are same we are not continuing to the code that compares the cost calculation. The fix should be to modify the condition as follows:
keyinfo->user_defined_key_parts <= best_key_parts

After modifying the code we see the optimizer is selecting the correct plan.

* DEF0370542: Fixing a mtr test case failure related to the previous change

* DEF0370542: Enhancing the fix to consider cost when key parts are equal. First attempt to create a mtr for the issue. Reverting back the mtr test case in previous commit.

* DEF0370542: Adding sn_order_by_limit_optimize_level dynamic variable.

This variable will control the level of optimizations made based on the
values in the order-by clause. Current values include the following:

0 - Generated plans are not reconsidered based on the values in the limit clause.
1 - Generated plans are reconsidered but plans generated in limit are selected only
based on the cost. The keyparts in keys are not considered when selecting the plan.
2 - Default. The plans are reconsidered and plans are selected based on both the keyparts
of keys and the cost. This is the default behaviour of vanilla MariaDB 10.4.

* DEF0370542: Fixing few failing mtrs. These mtrs are mainly failing because of output
mismatches. Fixed them.

* DEF0370542: Adding mtrs to verify order_by_level feature. Add code to execute the
second phase if force index is used on the table.

* DEF0370542: Adding files missed in the last commit.

* DEF0370542: Adding defect number to the tests

* DEF0370542: Adding sn_order_by_limit_optimize_levels 3, 4, and 5.
	3 - (Experimental) Same as 0, except the second pass for order by
        is executed if the row threshold specified by sn_order_by_row_threshold exceeds
        the row estimate for the table.
        4 - (Experimental) Same as 0, except the second pass for order by
        is executed if the cost for the referral key differs from the cost
        for the key part analysis by a factor of sn_order_by_factor_threshold. Not implemented.
        5 - (Experimental) Same as 0, except the second pass for order by is
        executed if reference key's access type is ALL. The second pass only
        uses cost comparison.

For level 3 the record threshold can be specified using sn_order_by_row_threshold.
Note that level 4 logic is not implemented.
Modified mtrs to accomodate new parameters.

* DEF0370542: Fixing the description.

* DEF0370542: Fixing two failing mtrs due to description change.

* DEF0370542: Fixing the condition for order by thread level 3. If row estimate < sn_order_by_row_threshold then execute order by limit optimization, otherwise (row estimate >= sn_order_by_row_threshold) skip order by reconsideration step.

* DEF0370542: Fixing description for level 3 and correcting associated mtrs.

* DEF0370542: Correcting a comment

* DEF0370542: Did several changes in this commit.
 1. Added mtrs for level 3 and level 5
 2. Made level 5 the default order by level.
 3. Renamed sn_order_by_row_threshold to sn_order_by_limit_row_threshold
 4. Fix the descriptions for level 3, 4, and 5.

* DEF0370542: reverting back the default order by level back to 2 due to multiple mtr failures.

* DEF0370542: Fixing few mtr failures due to variable rename and default value rename.

* DEF0370542: Reverting cmake maintainer and removing an empty file and its references from a test.