PRB1628125: Fixing rename table error after incremental backup restore in a TSE instance; porting fix to 10.4.20 (#91)
* PRB1628125: Fixing rename table error after incremental backup restore in a TSE instance, porting fix to 10.4.20.

This is a direct port to 10.4.20 from the fix made to 10.2.43.
An MDEV is not yet assigned.

The Problem
==========
Consider the following steps:

Suppose there is an encrypted table "t"
Take a base backup of the database
Create a table like t and name it tmp_t
Rename t to z_t and tmp_t to t
Start an increment backup, while the increment backup is ongoing, execute
optimize table on t (make sure this is executed after copying t to the backup)
Restore the incremental backup and start the server.
When we try to rename z_t table to t_prime, we get the following error:
errno: 155 "The table does not exist in the storage engine"

The Cause
========
An incremental backup collect pages modified since a given LSN number.
The modified pages are written to the backup as .delta files.
Further, tables created/optimized while the backup is on-going on are
saved with a suffix such as ".new", ".ren".

The step 5 above creates a t.new file. These files are processed
at the begining of the apply log stage when preparing the incremental
backup. When t.new file is processed, mariabackup deletes the
existing t.ibd file from the base directory and copies t.new as t.ibd.
However, the new t.ibd file carries a tablespace id different from
the old t.ibd file.

To apply the delta, mariabackup finds the proper tablespace file
in "xb_delta_open_matching_space". Mariabackup fails to
find a matching tablespace for z_t (created at step 4 above) inside
"xb_delta_open_matching_space", because there isn't a matching tablespace
in the base backup by the name z_t and also there isn't a matching space
with z_t's tablespace id. At this point, mariabackup creates new tablespace
file. However, this newly created tablespace file does not contain
crypt data and it just have 4 allocated pages. If delta file for
z_t table is empty, the restore operation succeeds but when
we try to access z_t after restore it remains corrupted. If delta file
contains pages (i.e., z_t is not empty), then we will see mariabackup
throwing an error saying that delta file cannot be applied because it cannot
read page 4 from the ibd file.

The Fix
=====
When processing .new files, deleting original t.ibd file and replacing
it with t.new file is the main cause for the issue.

Therefore, in the fix, when processing t.new files we do not
remove original t.ibd file, instead we rename the existing t.ibd
file to xtrabackup_tmp#<space_id>.ibd.

With this, when mariabackup try to find a matching space for z_t.delta.ibd, it will
return the xtrabackup_tmp#<space_id>.ibd and will be renamed to z_t.ibd in the
base backup. Then, afterwards, mariabackup should be able to apply the delta
to the z_t.ibd.

* PRB1628125: Adding have_debug.inc to tests. Otherwise, these tests will fail in a release build.