DEF0315061: Fix incorrect binlog coordinates returned by backup restore
* Restore getting InnoDB position from mariabackup --no-lock

Revert the patch for MDEV-18917, which removed this functionality.
This restores that mariabackup --prepare recovers the transactional
binlog position from the redo log, and writes it to the file
xtrabackup_binlog_pos_innodb.

This position is updated only on every InnoDB commit. This means that
if the last event in the binlog at the time of backup is a DDL or
non-transactional update, the recovered position from --no-lock will
be behind the state of the backup.

Signed-off-by: Kristian Nielsen <knielsen@knielsen-hq.org>

* Test case for provisioning a slave with mariabackup --no-lock

Signed-off-by: Kristian Nielsen <knielsen@knielsen-hq.org>

* Revert: MDEV-22351 InnoDB may recover wrong information after RESET MASTER

This commit can cause the wrong (old) binlog position to be recovered by
mariabackup --prepare. It implements that the value of the FIL_PAGE_LSN is
compared to determine which binlog position is the last one and should be
recoved. However, it some cases it is seen that the FIL_PAGE_LSN order does
not match the binlog position commit order, as is assumed by the code.

In one example, the recovery first encountered this in trx_rseg_mem_restore():

  lsn=27282754  binlog position (./master-bin.000001, 472908)

and then later:

  lsn=27282699  binlog position (./master-bin.000001, 477164)

The last one 477164 is the correct position. However, because the LSN
encountered for the first one is higher, that position is recovered instead.
This results in too old binlog position, and a newly provisioned slave will
start replicating too early and get duplicate key error or similar.

Signed-off-by: Kristian Nielsen <knielsen@knielsen-hq.org>

DEF0315061: Replace $SERVER_MYPORT_1 in slave_provision_nolock.result