SAMU-157 Revert most of merge, leave only bug fix
Looks like merging the whole function was bad idea as it may depend on
more patches. Though the changes looked innocent on the first glance,
it turns out they are not good anyway. main, mariabackup suites didn't
show the difference to me though.

But now the fix is minimal, only missing condition is added. Sorry!