MDEV-19929 Add a startup message about PAUSE instruction timing
On IA-32 and AMD64 (x86-64) CPU implementations, the latency of the
PAUSE instruction (REPZ NOP that was repurposed from a no-operation
on the Intel Pentium 4) can vary a lot, and we may adjust the
my_cpu_relax_multiplier accordingly. Let us display the value of
that variable on startup if --log-warnings=3 or larger was specified.

Microsoft Windows is available multiple instruction set architectures.
Therefore, we had better not assume that defined(_WIN32) implies the
availability of a PAUSE instruction.