Generate a sequence of functions (execution plan) to call per each row.
With a growth of richness of mariadb capabilities, the overhead for each row
also grows.

In this patch, instead of making support checks for every row call, an array of
functions to be called is generated instead.

Thus, the most richless case will only call two pointers to members.

The abi is optimized for calling update_row, so the prototype is made hatching
it:
int (handler::*member)(const uchar*,const uchar*)

Every single operation (such as: triggers, view checks, long UNIQUE, WITHOUT
OVERLAPS checks, row logging) is wrapped into a handler method with this
prototype.

To pass extra parameters, static thread_local is used for now. The next patch
resolves it and puts all extra methods in the separate class.

Around 20 checks per row operation is removed in update with it.

The row execution is more complex then simply calling the array of funcitons:
sometimes we should skip the row update, or the whole row execution (as
defined by the view visibility).

To handle this, the array is marked up with labels during the generation.
The set of labels is pre-defined for the operation. Now it's represented
with struct handler::Update_execution_markup.

int result is supposed to match the following format:
-1 == EOF
0 == success
1 == generic error
2 == skip row (==VIEW_CHECK_SKIP)
3-... == skip to some label in the function array
>= HA_ERR_FIRST : handler error