MDEV-34062: Implement innodb_log_file_mmap on 64-bit systems
When using the default innodb_log_buffer_size=2m, mariadb-backup --backup
would spend a lot of time re-reading and re-parsing the log. For reads,
it would be beneficial to memory-map the entire ib_logfile0 to the
address space (typically 48 bits or 256 TiB) and read it from there,
both during --backup and --prepare.

We will introduce the Boolean read-only parameter innodb_log_file_mmap
that will be OFF by default on most platforms, to avoid aggressive
read-ahead of the entire ib_logfile0 in when only a tiny portion would be
accessed. On Linux and FreeBSD the default is innodb_log_file_mmap=ON,
because those platforms define a specific mmap(2) option for enabling
such read-ahead.

We also experimented with allowing writes of the ib_logfile0 via a
memory mapping. A fundamental problem with that would be unnecessary
read-before-write in case of a major page fault, that is, when a new,
not yet cached, virtual memory page in the circular ib_logfile0 is
being written to. There appears to be no way to tell the operating system
that we do not care about the previous contents of the page, or that the
page fault handler should just zero out the contents.

Most references to HAVE_PMEM or log_sys.is_pmem() are replaced with
references to HAVE_INNODB_MMAP or log_sys.is_mmap().

Memory-mapped regular files differ from log_sys.is_pmem() in the way that
an open file handle to ib_logfile0 will be retained. In both code paths,
log_sys.is_mmap() will hold. Holding a file handle open will allow
log_t::clear_mmap() to disable the interface with fewer operations.