MDEV-34791 btr_cur_t::search_leaf() is suboptimal for single-page indexes
btr_cur_t::search_leaf(): When the index root page is also a leaf page
and we need to acquire an exclusive latch on it and upgrade the existing
shared latch to exclusive. Thanks to the index tree latch and the
buffer fix that we will be holding even if we temporarily release
the page latch while waiting for an exclusive latch, the root page must
remain valid and the index tree must remain a single page.