MDEV-34750 SET GLOBAL innodb_log_file_size is not crash safe
The recent commit 4ca355d863e3b6a14439eebbb2958afccb3548e3 (MDEV-33894)
caused a serious regression for online InnoDB ib_logfile0 resizing,
breaking crash-safety unless the memory-mapped log file interface is
being used. However, the log resizing was broken also before this.

To prevent such regressions in the future, we will extend the test
innodb.log_file_size_online with a kill and restart of the server
and with some writes running concurrently with the log size change.

After this change, SET GLOBAL innodb_log_file_size may block
execution until there have been writes to persistent InnoDB tables.
This is because the log resizing can only be completed on log
checkpoint completion.

log_t::resize_start(): Do not allow the resized log to start before
the current log sequence number. In this way, there is no need to
copy anything to the first block of resize_buf. The previous logic
regarding that was incorrect in two ways. First, we would have to
copy from the last written buffer (buf or flush_buf). Second, we failed
to ensure that the mini-transaction end marker bytes would be 1
in the buffer.

log_t::resize_write_buf(): Take the buffer
(resize_buf or resize_flush_buf) as a parameter. We are not always
swapping buffers. Do not write anything that would be before the
log_sys.resize_lsn.

log_t::append(): Define as a static function; only some debug
assertions need to refer to the log_sys object.

innodb_log_file_size_update(): Wake up the buf_flush_page_cleaner()
if needed, and wait for it to complete a batch while waiting for
the log resizing to be completed. If the buf_pool.flush_list is
empty, our wait will time out in 5 seconds, so that we can
periodically check if the execution of SET GLOBAL innodb_log_file_size
was aborted.